requirements.txt
# SWE-bench Comparison Framework Requirements
# ============================================

# Core Dependencies
# -----------------
pydantic>=2.0.0
pyyaml>=6.0
tiktoken>=0.5.0
click>=8.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0
numpy>=1.24.0

# Data Processing
# ---------------
datasets>=2.14.0
pandas>=2.0.0
requests>=2.31.0

# Machine Learning & Embeddings
# -------------------------------
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4

# Retrieval & Search
# ------------------
rank-bm25>=0.2.2
scikit-learn>=1.3.0
scipy>=1.11.0

# Code Analysis
# -------------
tree-sitter>=0.20.0
tree-sitter-python>=0.20.0
jedi>=0.19.0
ast-decompiler>=0.7.0

# Evaluation Metrics
# ------------------
codebleu>=0.6.0
sacrebleu>=2.3.0
rouge-score>=0.1.2
nltk>=3.8.0

# Patch Generation
# ----------------
diff-match-patch>=20230430
unidiff>=0.7.0

# Docker Support (optional)
# -------------------------
docker>=6.1.0

# Experiment Tracking (optional)
# ------------------------------
mlflow>=2.6.0
# wandb>=0.15.0

# Visualization (optional)
# ------------------------
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Reporting
# ---------
jinja2>=3.1.0
markdown>=3.4.0
weasyprint>=59.0  # For PDF generation

# Development & Testing
# ---------------------
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Type Stubs
# ----------
types-pyyaml>=6.0.0
types-requests>=2.31.0
